# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Panier API Integration",
        "description": "API endpoints for communicating with the Panier API - ZIMRA Fiscalisation Support.\r\n\r\n## Rate Limit\r\n500 API calls every 5 minutes for each unique IP Address.\r\n\r\n## Authentication\r\nRequires APP-ID and API-KEY headers obtained from your Panier company account's Developer API section.\r\n\r\n## ZIMRA Fiscalisation\r\nThis API supports ZIMRA Fiscal Invoices, Credit Notes and Debit Notes based on ZIMRA Fiscal Device Gateway API Specs v7.2.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/invoice/create": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Create invoices",
                "description": "Create invoices with optional ZIMRA fiscalization",
                "operationId": "5bec6634887372b030fc6efba08cc512",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "zimra_fiscalize": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/invoice/update": {
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoices",
                "operationId": "924748caa6ecfb9e63ad75d35fca8cb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/invoice/search": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Search invoices",
                "operationId": "5401671fd0a544c8d56ff73f6652946a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/invoice/convert-to-sale": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Convert invoice to sale",
                "operationId": "e7151e995007142a7162ea8feccdd8ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "properties": {
                                            "invoice_id": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "zimra_fiscalize": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/invoice/download": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Download invoice as PDF",
                "operationId": "bcfc112dd01d6d1741563bafed9d817e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
        "/invoice/delete": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Delete invoices",
                "operationId": "322cd673a2b0ee39893612635e34c76e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "data"
                                ],
                                "properties": {
                                    "data": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id"
                                            ],
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "AppId": [],
                        "ApiKey": []
                    }
                ],
                "servers": [
                    {
                        "url": "/api/v1",
                        "description": "Local API Server"
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "AppId": {
                "type": "apiKey",
                "description": "Panier APP-ID from Developer API section",
                "name": "APP-ID",
                "in": "header"
            },
            "ApiKey": {
                "type": "apiKey",
                "description": "Panier API-KEY from Developer API section",
                "name": "API-KEY",
                "in": "header"
            }
        }
    },
...
}